requirements.txt
aiohttp
bson
fastapi
numpy
pydantic
pypdf
pymilvus
pymongo
python-dotenv
requests
schedule
selenium
sentence-transformers
torch
uvicorn
webdriver-manager
beautifulsoup4
flask